Move `std::io::pipe` code into its own file
